requirements.txt
pdf2image>=1.16.3
Pillow>=10.0.0
PyPDF2>=3.0.0
streamlit>=1.28.0
